Added test for comparing questions
